+++ src/main/java/com/quasiris/qsf/pipeline/filter/elastic/Profiles.java
@@ -6,4 +6,10 @@ public class Profiles {
     public static String matchAll() {
         return "classpath://com/quasiris/qsf/elastic/profiles/match-all-profile.json";
     }
+
+    public static String queryString() {
+        return "classpath://com/quasiris/qsf/elastic/profiles/query-string-profile.json";
+    }
+
+
 }